Working on help files..

+++ src/net/sf/mzmine/modules/rawdatamethods/filtering/datasetfilters/DataSetFilteringTask.java
@@ -16,35 +16,26 @@
  * MZmine 2; if not, write to the Free Software Foundation, Inc., 51 Franklin
  * St, Fifth Floor, Boston, MA 02110-1301 USA
  */
-package net.sf.mzmine.modules.rawdatamethods.filtering.scanfilters;
+package net.sf.mzmine.modules.rawdatamethods.filtering.datasetfilters;
 
-import java.io.IOException;
 import java.lang.reflect.Constructor;
-import java.util.logging.Logger;
 
 import net.sf.mzmine.data.ParameterSet;
 import net.sf.mzmine.data.RawDataFile;
-import net.sf.mzmine.data.RawDataFileWriter;
-import net.sf.mzmine.data.Scan;
 import net.sf.mzmine.data.impl.SimplePeakList;
 import net.sf.mzmine.main.MZmineCore;
-import net.sf.mzmine.modules.rawdatamethods.filtering.scanfilters.preview.RawDataFilter;
+import net.sf.mzmine.modules.rawdatamethods.filtering.datasetfilters.preview.RawDataFilter;
 import net.sf.mzmine.taskcontrol.Task;
 import net.sf.mzmine.taskcontrol.TaskStatus;
 
 /**
  * @see
  */
-class RawDataFilteringTask implements Task {
+class DataSetFilteringTask implements Task {
 
-	private Logger logger = Logger.getLogger(this.getClass().getName());
-	private RawDataFile dataFile,  filteredRawDataFile;
+	private RawDataFile[] dataFiles;
 	private TaskStatus status = TaskStatus.WAITING;
 	private String errorMessage;
-
-	// scan counter
-	private int processedScans = 0,  totalScans;
-	private int[] scanNumbers;
 
 	// User parameters
 	private String suffix;
@@ -57,13 +48,13 @@ class RawDataFilteringTask implements Task {
 	private SimplePeakList newPeakList;
 
 	/**
-	 * @param dataFile
+	 * @param dataFiles
 	 * @param parameters
 	 */
-	RawDataFilteringTask(RawDataFile dataFile,
-			RawDataFilteringParameters parameters) {
+	DataSetFilteringTask(RawDataFile[] dataFiles,
+			DataSetFiltersParameters parameters) {
 
-		this.dataFile = dataFile;
+		this.dataFiles = dataFiles;
 
 		rawDataFilterTypeNumber = parameters.getRawDataFilterTypeNumber();
 		this.parameters = parameters.getRawDataFilteringParameters(rawDataFilterTypeNumber);
@@ -76,17 +67,17 @@ class RawDataFilteringTask implements Task {
 	 * @see net.sf.mzmine.taskcontrol.Task#getTaskDescription()
 	 */
 	public String getTaskDescription() {
-		return "Filtering scans in " + dataFile;
+		return "Filtering raw data";
 	}
 
 	/**
 	 * @see net.sf.mzmine.taskcontrol.Task#getFinishedPercentage()
 	 */
 	public double getFinishedPercentage() {
-		if (totalScans == 0) {
-			return 0;
+		if (rawDataFilter != null) {
+			return rawDataFilter.getProgres();
 		} else {
-			return (double) processedScans / totalScans;
+			return 0;
 		}
 	}
 
@@ -104,10 +95,6 @@ class RawDataFilteringTask implements Task {
 		return errorMessage;
 	}
 
-	public RawDataFile getDataFile() {
-		return dataFile;
-	}
-
 	/**
 	 * @see net.sf.mzmine.taskcontrol.Task#cancel()
 	 */
@@ -122,13 +109,11 @@ class RawDataFilteringTask implements Task {
 
 		status = TaskStatus.PROCESSING;
 
-		logger.info("Started filtering scans on " + dataFile);
+		//logger.info("Started filtering scans on " + dataFile);
 
-		scanNumbers = dataFile.getScanNumbers(1);
-		totalScans = scanNumbers.length;
 
 		// Create raw data filter according with the user's selection
-		String rawDataFilterClassName = RawDataFilteringParameters.rawDataFilterClasses[rawDataFilterTypeNumber];
+		String rawDataFilterClassName = DataSetFiltersParameters.rawDataFilterClasses[rawDataFilterTypeNumber];
 		try {
 			Class rawDataFilterClass = Class.forName(rawDataFilterClassName);
 			Constructor rawDataFilterConstruct = rawDataFilterClass.getConstructors()[0];
@@ -139,48 +124,25 @@ class RawDataFilteringTask implements Task {
 			return;
 		}
 		try {
+			for (RawDataFile dataFile : dataFiles) {
+				RawDataFile newDataFiles = rawDataFilter.getNewDataFiles(dataFile);
+				if (newDataFiles != null) {
+                    String name = newDataFiles.getName().replace("-Filtered", "");
+					newDataFiles.setName(name + suffix);
+					MZmineCore.getCurrentProject().addFile(newDataFiles);
 
-			// Create new raw data file
+					// Remove the original file if requested
+					if (removeOriginal) {
+						MZmineCore.getCurrentProject().removeFile(dataFile);
 
-			String newName = dataFile.getName() + " " + suffix;
-			RawDataFileWriter rawDataFileWriter = MZmineCore.createNewFile(newName);
-
-			for (int i = 0; i < totalScans; i++) {
-
-				if (status == TaskStatus.CANCELED) {
-					return;
+					}
 				}
-
-				Scan scan = dataFile.getScan(scanNumbers[i]);
-				if ((scan.getMSLevel() != 1) && (scan.getParentScanNumber() <= 0)) {
-					return;
-				}
-				Scan newScan = rawDataFilter.getNewScan(scan);
-				if (newScan != null) {
-					rawDataFileWriter.addScan(newScan);
-				}
-
-				processedScans++;
-			}
-
-
-			// Finalize writing
-			try {
-				filteredRawDataFile = rawDataFileWriter.finishWriting();
-				MZmineCore.getCurrentProject().addFile(filteredRawDataFile);
-
-
-				// Remove the original file if requested
-				if (removeOriginal) {
-					MZmineCore.getCurrentProject().removeFile(dataFile);
-				}
-			} catch (Exception exception) {
 			}
 
 			status = TaskStatus.FINISHED;
-			logger.info("Finished scan filter on " + dataFile);
+		//logger.info("Finished scan filter on " + dataFile);
 
-		} catch (IOException e) {
+		} catch (Exception e) {
 			status = TaskStatus.ERROR;
 			errorMessage = e.toString();
 			return;
